Revert "Refactor code"
